Rename page_streaming to paged_iteration (#99)
We are going to rename this due to avoid naming of 'streaming',
the last remaining reference is the file name. This does not
change anything in the code.

This also splits lib/api_callable to multiple files -- that file
contained many independent test sections. I think the newer
file structure is clearer and easier to edit.