test: add mock server for testing (#801)
* test: add mock server for testing

Adds a mock Spanner server for testing purposes. The mock server implements
the gRPC interface of the Spanner backend and implements simple in-mem
functionality for creating sessions and transactions. The results of query
and update statements must be mocked in advance before they can be executed
on the server. The server does not implement any storage, and executing an
update statement on the server does not change the state of the server.

The purpose of the server is to allow simple in-mem (integration) testing
for specific use cases and error conditions.

* fix: add line endings

* fix: process review comments