feat: return ResultSetMetadata for query (#1308)
* feat: return ResultSetMetadata for query

Returns the ResultSetMetadata for queries. This makes it easier to
create a generic view of the results of a query.

This change also adds the option of returning the values of nameless
columns in JSON results. These were previously always filtered away.
Setting the option `includeNameless` in JSONOptions will now include
these values as well.

Fixes #1307

* fix: remove commented code