samples: add schema samples, typescript sample support (#1262)
fixes: https://github.com/googleapis/nodejs-pubsub/issues/1220

The remainder of this PR is about adding schema sample snippets, as well as some preliminary TypeScript snippet support. There are still some linting issues with the generated samples, and I wanted to talk about the method used here before committing to it, so it's still marked as a work in progress.