fix: updated typescript version for compiler (#901)
When running the generator locally, bazel demands the typescript version to be `">=3.0.0 <4.3` which conflicts with the typescript version of `"^4.3.2` the compiler is currently using. 

Therefore, this change is to downgrade the typescript version used for compiling the generator locally in order for bazel rules to work. The generator does not actually use any 4.3+ features, so this downgrade should not affect any functionality (and does not affect the typescript version used for the generated client libraries, which is the latest one). 

We can bump this version back up once bazel is able to accept the latest version of typescript.