chore(testproxy): Change config so test runner can work with proxy (#1245)
* Change config so test runner can work with proxy

We adjust this to allow the test proxy to use new changes in the test runner.

* Add the new proto for communicating with the proxy

We add a new proto and make adjustments to the test proxy to use the new proto. This change should enable us to work with the test runner properly.

* Remove the v2 test proxy

Remove the v2 test proxy because it is not used anymore. If we need it again we can just search for it in version control.