feat: introduce EntityFilter class with support for and/or filters (#1061)
* or filters interface

* implementation of new filter

* Remove duplicate code

* Parse with type guard

* Remove the newFilter variable

* Add filter and enable chaining

* test add filter

* Add another unit test

* Add system-test stubs

* Add system-tests for the OR filter

* Move things around

* Added a unit test

* Add a unit test for OR filters

* Just use filter method

* new warning test

* Revert "new warning test"

This reverts commit 37400a6d3a9d7385a0203db70949f037e4c35de4.

* Now removes deprecation warning properly

* Add a test for new warning

* Add setAncestor

Adds a new setAncestor method for ensuring only one ancestor is set for the query at a time. This will avoid errors that result because of setting multiple ancestors. Also deprecate hasAncestor because it will lead to warnings like this. Add parser logic to use the value provided in setAncestor for query sent to backend.

* Basic unit tests for setAncestor

Added tests for the query proto, one to make sure the query structure is right when setting ancestor once and one to make sure the query structure is right when setting ancestor twice. Also added a unit test to make sure that ancestor is set the right way internally when using setAncestor.

* change expected result for OR query

This code change adjusts the expected result for running an OR query. Old result used to correspond with AND, but now corresponds to OR.

* Fix a test by not requiring the done callback

A test is timing out because we are waiting for done to be called. This fix does not require done to be called.

* Revert "Fix a test by not requiring the done callback"

This reverts commit 1159b374f483d107736a765b5b4f1ade80116ce3.

* Revert "Basic unit tests for setAncestor"

This reverts commit 86841d6e0a3d26e2df647fa4a42ed849b61b38a6.

* Revert "Add setAncestor"

This reverts commit e84582faa6041429a4e81d47bd21dbc338a639e7.

* Separate filters and new filters internally

This commit is done to avoid a breaking typescript change which could have the potential to affect some users who read `filters` on a query as its type had been changed to be more flexible, but is now back to what it was.

* Move AND/OR into their own separate function

AND and OR should not be static functions of the filter class because then the user has to type Filter.AND instead of AND for example.

* Eliminate unused imports

Artifacts of having imports laying around and moving functionality between files

* Revert "Add a test for new warning"

This reverts commit bc15f325d87b4ffe46ae8727293c5cefa2300501.

* Revert "Now removes deprecation warning properly"

This reverts commit db02a5089242b25b4bebae4304a4778f0bf6d17f.

* Modify test cases to capture nuances in data

We add additional asserts to the data in order to capture the nuances of the composite operator. For example, for the OR test we make sure the filter doesn’t always require both conditions to be true.

* Added comments to code that was refactored

Code for building the `filter` property of the query proto was pulled into the `Filter` object. Comments indicate how legacy functionality was maintained and which lines of code perform which task.

* rename NewFilter to entity filter

rename new filter to entity filter to eliminate the need for an internal rename that causes confusion

* Switch around last and first

These test cases have mistakes in their names. We should change them to reflect the position of `hasAncestor`.

* Change the comment to reflect the new name

The name of the base class is now EntityFilter. Adjust this comment so that it correctly matches the parameter type.

* rename and move constant map up

rename composite filter functions so that they don’t look like constants and move a map up so that it doesn’t have to be initialized every time.

* Rename newFilters to entityFilters