fix: Allow user to set custom endpoints (#1101)
* Write a test for custom endpoints

With the code in its present state, this test should fail. A fix for regional endpoints intends to clean this up.

* Limit setting credentials to insecure

Only set the credentials to insecure if the custom endpoint is localhost and credentials are not provided. This way the user can set custom endpoints with the client and the request will work.

* Change the tests for custom endpoints

There is a test that uses a mock which is preventing the base url from being set. The test should be more realistic and when it is, the test passes with the new change.

* Add another localhost option

There is a third way the user can specify localhost. We should include this way as an option.