build: integrate nodejs core libraries (#848)
* build: run interdependent tests

* build: run interdependent tests

* fix

* fix(ci): make run-interdependent-tests.sh safer

The script was failing when running in a directory that does not have a 'packages' subdirectory. This commit adds a check to ensure the directory exists before trying to list its contents.

* fix(deps): syncpack dependency versions

This commit fixes the dependency mismatches reported by syncpack.

* fix(ts): address type errors from @types/node update

This commit fixes the TypeScript errors that were introduced after updating the dependencies with syncpack. The errors were related to the new definition of Uint8Array in @types/node.

* ci(syncpack): automate fixing of mismatches

This commit updates the syncpack workflow to automatically fix dependency mismatches and commit the changes.

* Update package.json

* revert(ci): remove automated fix for syncpack

This reverts the previous commit that automated the fixing of syncpack mismatches.

* fix(deps): update dependencies to resolve syncpack mismatches

* fix(proto3-json-serializer): resolve Uint8Array and Buffer type issues

* run lint

* feat: add top-level lint-fix script

This commit adds a top-level lint-fix script that runs  in all packages. It also fixes an issue in  package that was causing the linting to fail.

* fix(lint): run gts fix on all packages

* chore: make sure we are actually installing packages locally

* remove the package-lock.json

* fix

* chore: fix

* retry

* chore: fix

* Update continuous.yaml

* run lint

* chore: fix nodejs runtime

* address comments

* remove numbers

* pin typescript dependency

---------

Co-authored-by: miguel <miguelvelezsa@google.com>